# Conflicts:
#	backend/src/controllers/cityController.ts
#	frontend/src/backOffice/components/common/AsideMenu/AsideMenu.tsx
#	frontend/src/backOffice/pages/administrator/Cities/Cities.tsx

+++ backend/src/controllers/cityController.ts
@@ -12,30 +12,29 @@ import { User, UserRole } from "../entities/User";
 export const CityController: IController = {
   // GET ALL CITIES
 
-  getCities: async (req: Request, res: Response): Promise<void> => {
-    try {
-      const allCities = await dataSource.getRepository(City).find({
-        relations: {
-          userAdminCity: true,
-          poi: true,
-        },
-        select: {
-          userAdminCity: {
-            id: true,
-            username: true,
-            email: true,
-            city: true,
-            image: true,
-            createdPoi: true,
-          },
-        },
-      });
-      res.status(200).send(allCities);
-    } catch (err) {
-      console.log(err);
-      res.status(400).send({ error: "Error while reading cities" });
-    }
-  },
+	getCities: async (req: Request, res: Response): Promise<void> => {
+		try {
+			const allCities = await dataSource.getRepository(City).find({
+				relations: {
+					userAdminCity: true,
+					poi: true,
+				},
+				select: {
+					userAdminCity: {
+						id: true,
+						username: true,
+						email: true,
+						city: true,
+						image: true,
+						createdPoi: true,
+					},
+				},
+			});
+			res.status(200).send(allCities);
+		} catch (err) {
+			res.status(400).send({ error: "Error while reading cities" });
+		}
+	},
 
   // GET ONE CITY BY ID
 
@@ -74,15 +73,20 @@ export const CityController: IController = {
 
   // CREATE CITY
 
-  createCity: async (req: Request, res: Response): Promise<void> => {
-    try {
-      const { name, image, coordinates, userAdminCityId } = req.body;
-      const { userId } = req.params;
+	createCity: async (req: Request, res: Response): Promise<void> => {
+		const { name, coordinates, userAdminCity } = req.body;
+		const { userId } = req.params;
 
-      // Check if current user is admin
-      const currentUser = await dataSource
-        .getRepository(User)
-        .findOne({ where: { id: userId } });
+		req.body.coordinates = [
+			parseInt(req.body.coordinates.split(",")[0], 10),
+			parseInt(req.body.coordinates.split(",")[1], 10),
+		];
+
+		try {
+			// Check if current user is admin
+			const currentUser = await dataSource
+				.getRepository(User)
+				.findOne({ where: { id: userId } });
 
       if (currentUser?.role !== UserRole.ADMIN) {
         res.status(403).send({
@@ -93,43 +97,77 @@ export const CityController: IController = {
         return;
       }
 
-      // check if name is alpha or not empty
-      // isAlpha check if string because only letters
-      if (
-        !validator.matches(
-          name,
-          /^[a-zA-ZàáâäãåąčćęèéêëėįìíîïłńòóôöõøùúûüųūÿýżźñçčšžÀÁÂÄÃÅĄĆČĖĘÈÉÊËÌÍÎÏĮŁŃÒÓÔÖÕØÙÚÛÜŲŪŸÝŻŹÑßÇŒÆČŠŽ∂ð '-]{2,100}$/
-        ) ||
-        validator.isEmpty(name, { ignore_whitespace: true })
-      ) {
-        res.status(400).send({
-          error: `Field must contains only characters (min: 2, max: 100)`,
-        });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-        return;
-      }
+			// Get UserAdminCityId
+			const currentUserAdminCity = await dataSource
+				.getRepository(User)
+				.findOne({
+					where: { username: userAdminCity },
+				});
+
+			const currentUserAdminCityBody = {
+				id: currentUserAdminCity?.id,
+				username: currentUserAdminCity?.username,
+				email: currentUserAdminCity?.email,
+				password: currentUserAdminCity?.password,
+				image: currentUserAdminCity?.image,
+				role: UserRole.ADMIN_CITY,
+				city: currentUserAdminCity?.city,
+			};
+
+			if (currentUserAdminCity === null) {
+				res.status(403).send({
+					error: "User admin city doesn't exist",
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+				return;
+			}
+
+			// check if name is alpha or not empty
+			// isAlpha check if string because only letters
+			if (
+				!validator.matches(
+					name,
+					/^[a-zA-ZàáâäãåąčćęèéêëėįìíîïłńòóôöõøùúûüųūÿýżźñçčšžÀÁÂÄÃÅĄĆČĖĘÈÉÊËÌÍÎÏĮŁŃÒÓÔÖÕØÙÚÛÜŲŪŸÝŻŹÑßÇŒÆČŠŽ∂ð '-]{2,100}$/
+				) ||
+				validator.isEmpty(name, { ignore_whitespace: true })
+			) {
+				res.status(400).send({
+					error: `Field must contains only characters (min: 2, max: 100)`,
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+				return;
+			}
 
       // check if coordinates are type [number, number]
 
-      if (coordinates.length > 2) {
-        res.status(400).send({
-          error: "Incorrect format of coordinates (must be [lat, long])",
-        });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-        return;
-      }
+			if (
+				req.body.coordinates !== null &&
+				(req.body.coordinates.length > 2 ||
+					(typeof req.body.coordinates[0] !== "number" &&
+						typeof req.body.coordinates[1] !== "number"))
+			) {
+				res.status(400).send({
+					error: "Incorrect format of coordinates (must be [lat, long])",
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+				return;
+			}
 
-      // check if userAdminCity is UUID type
-      if (userAdminCityId !== null && !validator.isUUID(userAdminCityId)) {
-        res.status(400).send({
-          error: "Incorrect format of admin city id (must be uuid)",
-        });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-        return;
-      }
+			// check if userAdminCity is UUID type -
+			if (
+				currentUserAdminCity !== null &&
+				!validator.isUUID(currentUserAdminCity.id)
+			) {
+				res.status(400).send({
+					error: "Incorrect format of admin city id (must be uuid)",
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+				return;
+			}
 
       // check if name doesn't already exist in db
       const nameAlreadyExist = await dataSource
@@ -154,47 +192,53 @@ export const CityController: IController = {
         return;
       }
 
-      // rename file image
-      if (req.file !== undefined) {
-        const originalname = req.file.originalname;
-        const filename = req.file.filename;
-        const newName = `${uuidv4()}-${originalname}`;
-        fs.rename(
-          `./public/city/${filename}`,
-          `./public/city/${newName}`,
-          (err) => {
-            if (err !== null) throw err;
-          }
-        );
-        req.body.image = `/public/city/${newName}`;
-      } else {
-        res.status(400).send({ error: "An image is required" });
-        return;
-      }
+			// rename file image
 
-      // format coordinates
-      req.body.coordinates = {
-        type: "Point",
-        coordinates: [coordinates[0], coordinates[1]],
-      };
+			if (req.file !== undefined) {
+				const originalname = req.file.originalname;
+				const filename = req.file.filename;
+				const newName = `${uuidv4()}-${originalname}`;
+				fs.rename(
+					`./public/city/${filename}`,
+					`./public/city/${newName}`,
+					(err) => {
+						if (err !== null) throw err;
+					}
+				);
+				req.body.image = `/public/city/${newName}`;
+			} else {
+				res.status(400).send({ error: "An image is required" });
+				return;
+			}
 
-      await dataSource.getRepository(City).save(req.body);
-      res.status(201).send("Created city");
-    } catch (error: any) {
-      // check if error is 'Key ("userAdminCityId")=(id) already exists'
-      if (error.code === "23505") {
-        res.status(409).send({
-          error: "User is already administrator in another city",
-        });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-      } else {
-        res.status(400).send({ error: error.message });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-      }
-    }
-  },
+			// format coordinates
+			req.body.coordinates = {
+				type: "Point",
+				coordinates: [req.body.coordinates[0], req.body.coordinates[1]],
+			};
+
+			await dataSource
+				.getRepository(City)
+				.save({ ...req.body, userAdminCity: currentUserAdminCity });
+			await dataSource
+				.getRepository(User)
+				.update(currentUserAdminCity.id, currentUserAdminCityBody);
+			res.status(201).json("Created city");
+		} catch (error: any) {
+			// check if error is 'Key ("userAdminCityId")=(id) already exists'
+			if (error.code === "23505") {
+				res.status(409).send({
+					error: "User is already administrator in another city",
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+			} else {
+				res.status(400).send({ error: error.message });
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+			}
+		}
+	},
 
   // UPDATE CITY BY ID
 
@@ -205,12 +249,9 @@ export const CityController: IController = {
     const { userId } = req.params;
 
 		req.body.coordinates = [
-			parseInt(coordinates[0], 10),
-			parseInt(coordinates[1], 10),
+			parseInt(req.body.coordinates.split(",")[0], 10),
+			parseInt(req.body.coordinates.split(",")[1], 10),
 		];
-
-		console.log("req.body", req.body);
-		console.log("req.params", req.params);
 
 		try {
 			// Check if current user is admin
@@ -276,14 +317,18 @@ export const CityController: IController = {
 
       // check if image is alpha or not empty
 
-      if (image !== null && validator.isEmpty(image)) {
-        res.status(400).send({
-          error: `Field must contains only characters`,
-        });
-        if (req.file !== undefined)
-          await unlink(`./public/city/${req.file?.filename}`);
-        return;
-      }
+			if (
+				image !== null &&
+				typeof image === "string" &&
+				validator.isEmpty(image)
+			) {
+				res.status(400).send({
+					error: `Field must contains only characters`,
+				});
+				if (req.file !== undefined)
+					await unlink(`./public/city/${req.file?.filename}`);
+				return;
+			}
 
 			// check if coordinates are type [number, number]
 			if (
@@ -382,18 +427,17 @@ export const CityController: IController = {
         );
         req.body.image = `/public/city/${newName}`;
 
-        // delete the older file
-        if (cityToUpdate.image !== null) {
-          await unlink("." + cityToUpdate.image);
-        }
-      }
+				// delete the older file
+				if (cityToUpdate.image?.includes("public/city")) {
+					await unlink("." + cityToUpdate.image);
+				}
+			}
 
 			await dataSource.getRepository(City).update(id, req.body);
 			await dataSource
 				.getRepository(User)
 				.update(currentUserAdminCity.id, currentUserAdminCityBody);
-
-			res.status(200).send("Updated city");
+			res.status(200).json("Updated city");
 		} catch (error: any) {
 			// check if error is 'Key ("userAdminCityId")=(id) already exists'
 			if (error.code === "23505") {
@@ -426,19 +470,39 @@ export const CityController: IController = {
           error: "You are not authorized to delete a city",
         });
 
-        return;
-      }
-      // check if city exists in db
-      const cityToDelete = await dataSource
-        .getRepository(City)
-        .findOneBy({ id });
-      if (cityToDelete === null) {
-        res.status(404).send({ error: "City not found" });
-        return;
-      }
+				return;
+			}
 
-			await dataSource.getRepository(City).delete(id);
-			console.log(cityToDelete.image.includes("loremflickr"));
+			// check if city exists in db
+			const cityToDelete = await dataSource.getRepository(City).findOne({
+				where: { id },
+				relations: { userAdminCity: true },
+				select: {
+					userAdminCity: {
+						id: true,
+					},
+				},
+			});
+
+			if (cityToDelete === null) {
+				res.status(404).send({ error: "City not found" });
+				return;
+			}
+
+			// remove status admin to user
+			const userAdminCity = await dataSource
+				.getRepository(User)
+				.findOne({ where: { id: cityToDelete.userAdminCity.id } });
+
+			if (userAdminCity !== null) {
+				delete (userAdminCity as any).favouritePoi;
+				delete (userAdminCity as any).favouriteCities;
+				userAdminCity.role = UserRole.FREE_USER;
+				await dataSource.getRepository(City).delete(id);
+				await dataSource
+					.getRepository(User)
+					.update(userAdminCity.id, userAdminCity);
+			}
 
 			if (
 				cityToDelete.image !== null &&
@@ -448,8 +512,6 @@ export const CityController: IController = {
 			}
 			res.status(200).send("Deleted city");
 		} catch (err) {
-			console.log("err", err);
-
 			res.status(400).send({ error: "Error while deleting city" });
 		}
 	},
